requirements.txt
asgiref
crispy-bootstrap5
distlib
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-email
django-environ
django-filter
djangorestframework
filelock
keyboard
Markdown
MouseInfo
numpy
opencv-python
pillow
platformdirs
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
pytweening
sqlparse
typing_extensions
tzdata
virtualenv